[resolves #3] Implement Ko rule
This rule is implemented as an addition to the `Board`, `LastTaker`,
which specifies the `Stone`, if any, that was placed in the last
turn that caused a capture of a `Stone` from the other player.
If the next turn causes a capture of that `LastTaker`, then
a Ko violation error is returned.