[resolves #26] Add a link to the GraphiQL browser on the homepage (#39)
* Add a link to the GraphiQL browser on the homepage

Currently when loading the server in development an empty 404 page is
rendered. This is confusing to new developers as they are unsure if
the server is working properly, and if it is working properly, there
is no guidance as to what to do with the server.

* Parse duration as hours

This was causing the app to crash on boot. This change seems to work
for me locally. I'm open to other ideas to get a duration of a week
if there is a better way. :)

    $ go build
    $ ./go_stop
    could not parse auth key durationtime: unknown unit  week in duration 1 week

* Update test script to test new packages

* Do not link to GraphiQL if it is not currently enabled

This conditionally adds the link to the GraphiQL routes for developers
but it will not display when GraphiQL is disabled.

* GraphiQL is enabled in the development environment

* Remove server.go, it was moved to server/ in another branch

* GraphiQL is enabled by configuration

Always link to the `/graphql` route regardless of the configuration
value.